backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-multipart
python-dotenv
psycopg2-binary
pillow
pdf2image
pydantic
aiofiles
httpx
